requirements.txt
flask
google-generativeai
langchain
faiss
